Update some feature and add manage donhang
Update some feature and add manage donhang

+++ e-mobile-shop/e-mobile-shop/Models/ChiTietDonHang.cs
@@ -8,7 +8,7 @@ namespace e_mobile_shop.Models
         public string MaDh { get; set; }
         public string MaSp { get; set; }
         public int? SoLuong { get; set; }
-        public decimal? ThanhTien { get; set; }
+        public double? ThanhTien { get; set; }
         public string MaCtdh { get; set; }
 
         public virtual DonHang MaDhNavigation { get; set; }
